Move deep-partial type to common-types
