+++ src/lib/common-types.ts
@@ -96,3 +96,8 @@ export type ValidateActionResponse = {
   statusCode?: number;
   statusMessage?: string;
 };
+
+// DeepPartial utility type for deep optional properties
+export type DeepPartial<T> = {
+  [P in keyof T]?: T[P] extends object ? DeepPartial<T[P]> : T[P];
+};
