+++ src/lib/mutate-request.ts
@@ -5,7 +5,7 @@ import { Operation } from "./enums";
 import { KubernetesObject } from "kubernetes-fluent-client";
 import { clone, mergeDeepRight } from "ramda";
 import { AdmissionRequest } from "./common-types";
-import { DeepPartial } from "./utility-types";
+import { DeepPartial } from "./common-types";
 
 // PeprMutateRequest class for mutation request handling
 export class PeprMutateRequest<T extends KubernetesObject> {
