--- src/lib/utility-types.ts
@@ -1,4 +1,0 @@
-// DeepPartial utility type for deep optional properties
-export type DeepPartial<T> = {
-  [P in keyof T]?: T[P] extends object ? DeepPartial<T[P]> : T[P];
-};
